requirements.txt
openai-whisper
sounddevice
numpy
pyperclip
keyboard
pystray
pillow
